halfway through lab
