+++ .env.sample
@@ -1,0 +1,2 @@
+DATABASE_URL=postgres://localhost:5432/basic-auth?sslmode=disable
+PORT=3002
