+++ src/auth/signin.js
@@ -1,0 +1,1 @@
+'use strict';
