+++ src/error-handlers/500.js
@@ -1,0 +1,11 @@
+'use strict';
+
+module.exports = (error, req, res, next) => {
+    res.status(500).send({
+        error: 500,
+        route: req.path,
+        query: req.query,
+        body: req.body,
+        message: `SERVER ERROR: ${error.message}`,
+    });
+};
