+++ src/models/user.schema.js
@@ -1,0 +1,12 @@
+'use strict';
+
+module.exports = sequelize.define('User', {
+    username: {
+        type: DataTypes.STRING,
+        allowNull: false,
+    },
+    password: {
+        type: DataTypes.STRING,
+        allowNull: false,
+    }
+});
